# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TAMAZIRT API Documentation",
        "description": "API for TAMAZIRT",
        "contact": {
            "email": "support@tamazirt.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "TAMAZIRT API Server"
        }
    ],
    "paths": {
...
        "/api/navigator/requests": {
            "get": {
                "tags": [
                    "Navigator Requests"
                ],
                "summary": "Get all assigned requests (Navigator only)",
                "description": "Retrieves a list of all service requests where the authenticated user is the assigned navigator. Requires 'navigator' role.",
                "operationId": "navigatorIndexRequests",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "Pending"
                                                    },
                                                    "tourist": {
                                                        "description": "Associated tourist details",
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized: navigator access required or not authenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/navigator/requests/{id}/status": {
            "patch": {
                "tags": [
                    "Navigator Requests"
                ],
                "summary": "Update request status (Navigator only)",
                "description": "Allows the navigator to update the status of an assigned request to Pending, Confirmed, or HandOff. Requires 'navigator' role and ownership of the request.",
                "operationId": "navigatorUpdateRequestStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the request to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "requestBody": {
                    "description": "New status for the request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Pending",
                                            "Confirmed",
                                            "HandOff"
                                        ],
                                        "example": "Confirmed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Request status updated successfully."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized: not the assigned navigator"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/navigator/requests/{id}": {
            "delete": {
                "tags": [
                    "Navigator Requests"
                ],
                "summary": "Delete an assigned request (Navigator only)",
                "description": "Deletes an assigned service request. Requires 'navigator' role and ownership of the request.",
                "operationId": "navigatorDeleteRequest",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the request to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Request deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized: not the assigned navigator"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Bearer Token Authentication",
                "in": "header",
                "bearerFormat": "Sanctum Token",
                "scheme": "Bearer"
            }
        }
    },
...
}